Safari-609.1.11.1/LayoutTests/fast/scrolling/ios/scroll-events-unresponsive-expected.txt

PASS receivedScrollEvent is false
PASS successfullyParsed is true

TEST COMPLETE
